requirements.txt
PySide2
Pillow
qrcode
